src/observer/requirements.txt
fastapi
uvicorn
PyYAML
types-PyYAML
